Removed additional info from Macros and fixed all extra features that will be added after MVP
