+++ Server/Application/Application.csproj
@@ -1,6 +1,23 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
     <ItemGroup>
+      <Compile Remove="ActivitiesFactor\**" />
+      <Compile Remove="Genders\**" />
+      <Compile Remove="Goals\**" />
+      <Compile Remove="HeightUnits\**" />
+      <Compile Remove="WeightUnits\**" />
+      <EmbeddedResource Remove="ActivitiesFactor\**" />
+      <EmbeddedResource Remove="Genders\**" />
+      <EmbeddedResource Remove="Goals\**" />
+      <EmbeddedResource Remove="HeightUnits\**" />
+      <EmbeddedResource Remove="WeightUnits\**" />
+      <None Remove="ActivitiesFactor\**" />
+      <None Remove="Genders\**" />
+      <None Remove="Goals\**" />
+      <None Remove="HeightUnits\**" />
+      <None Remove="WeightUnits\**" />
+    </ItemGroup>
+    <ItemGroup>
         <ProjectReference Include="..\Domain\Domain.csproj" />
         <ProjectReference Include="..\Persistence\Persistence.csproj" />
     </ItemGroup>
@@ -10,17 +27,11 @@
       <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="7.0.0" />
     </ItemGroup>
     <ItemGroup>
-      <Folder Include="Goals\" />
       <Folder Include="Errors\" />
-      <Folder Include="ActivitiesFactor\" />
       <Folder Include="Reference\" />
-      <Folder Include="Genders\" />
       <Folder Include="Users\" />
       <Folder Include="Interfaces\" />
       <Folder Include="Validation\" />
-      <Folder Include="HeightUnits\" />
-      <Folder Include="WeightUnits\" />
-      <Folder Include="UsersMacros\" />
     </ItemGroup>
     <PropertyGroup>
         <TargetFramework>netstandard2.1</TargetFramework>
