+++ Server/Application/MainDTOs/MealDto.cs
@@ -1,5 +1,4 @@
-﻿using Domain.Common;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Text;
 
@@ -8,9 +7,9 @@ namespace Application.MainDTOs
     public class MealDto
     {
         public int Id { get; set; }
+        public string Title { get; set; }
         public string Description { get; set; }
 
-        public Macros Macros { get; set; }
 
     }
 }
