--- Server/Application/MainDTOs/UserMacrosDto.cs
@@ -1,8 +1,0 @@
-﻿using System;
-namespace Application.MainDTOs
-{
-    public class UserMacrosDto
-    {
-        public int TotalMacros { get; set; }
-    }
-}
