+++ Server/Application/Users/Login.cs
@@ -5,7 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Application.Interfaces;
 using Domain;
-using Domain.Common;
+using Domain.User;
 using FluentValidation;
 using MediatR;
 using Microsoft.AspNetCore.Identity;
