--- Server/Domain/Common/UserMacrosType.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Domain.Common
-{
-    class UserMacrosType
-    {
-        public int Id { get; set; }
-        public Macros Macros { get; set; }
-        public bool IsRecommended { get; set; }
-    }
-}
