+++ Server/Infrastructure/Security/JwtGenerator.cs
@@ -5,7 +5,7 @@ using System.Security.Claims;
 using System.Text;
 using Application.Interfaces;
 using Domain;
-using Domain.Common;
+using Domain.User;
 using Microsoft.Extensions.Configuration;
 using Microsoft.IdentityModel.Tokens;
